Add files via upload
Had weird issues in first version of Validator.  Re did part of the original design and used the "break" statement in a few key spots instead of double truth condition while statements and understood how to truly initialize an array of Java objects (use a separate function devoted to making a single object, then invoke that repeatedly in a loop).